requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Faker==38.2.0
fastapi==0.122.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.12.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
starlette==0.50.0
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
